backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
python-docx==1.1.2
PyPDF2==3.0.1
openai==1.66.3
python-dotenv==1.0.1
typing-extensions==4.12.2

# Testing dependencies
pytest==8.0.2
httpx==0.28.1  # Required for TestClient
pytest-asyncio==0.23.5 